Change compare logic to avoid blind updates fix #7
Before, if the only difference we had was MAL last updated being older
than HB last update, then we performed a "blind update" meaning that we
still updated despite not having a specific value to change.

The benefit is that with this method we can also transfer values that
MAL API does not give (like comments) and thus we have no ability to
know if they have changed.

The drawback is that we are forced to do 2 updates in a row for each
anime (one with a fake episode change and one normal). That's because
MAL only updates an anime's last update if an episode changes. This lead
us with the case of updating, for example status (but not episode) which
in turn did not trigger a my_last_updated value change from MAL for that
anime. Thus the next time we checked, we would still see that specific
anime needing an update.

With this change, we mark the anime that only have a LastUpdated diff as
"uncertain" (they can appear with a more user friendly tag like "Okay"
in the web interface). The next time we sync, those "uncertain" anime
won't try to be updated. This will save us from the hacky method of
performing 2 updates (one of which was fake) to force MAL to update the
LastUpdated value and thus show the anime as "Updated" in the web
interface the next time we checked.

Now we will only update on values that we are absolutely sure that need
update. The drawback is that we will have one extra tag "Okay" on the
web interface which might require explanation to the users and in
general might complicate things.