Delete onepic package
The onepic package was simply fetching the first picture found using a
search term. It was used to display pictures of anime titles during
tests to get a better feel of the interface.

Google deprecated their image search API which was simple and
straightforward. For the same functionality we now have to use the
custom search API which requires making a specific custom search engine,
enable the anisync application to use the custom search API and use a
developer API key for each request. It makes things way too complex for
the use case of the onepic package which was only used for testing
anyways.